+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix positioning of Tooltip and Drop components if target is inline element
 - Fix Sticky component resizing on `clsBelow` class causes stutter
 - Fix Sticky component correctly resets on becoming inactive
+- Fix detection of finite mode in Slider component with `center` option enabled
 
 ## 3.15.17 (December 6, 2022)